Create 73. Set Matrix Zeroes.python
Intuition:
We need to modify the matrix in-place, so we can't use an auxiliary matrix or hash table. We can instead use the first row and first column of the original matrix as a replacement for the auxiliary array. This way, we can save the extra space required for the auxiliary arrays and also set the values in the first row and column to zero if any element in the corresponding row or column is zero.


Approach:
First, the code initializes two dummy vectors, Row and Col, with initial values of -1. These vectors will be used to mark the rows and columns that need to be set to zero.
The code then iterates through each element of the matrix and checks if it is zero. If an element is zero, it updates the corresponding indices in Row and Col to 0.
After marking the rows and columns, the code iterates through the matrix again. For each element, it checks if the corresponding row index or column index in Row or Col is zero. If either of them is zero, it sets the current element to zero.
Finally, the matrix will have rows and columns set to zero based on the values in Row and Col.